[5.4] remove trailing slash (#45566)
This pr removes the trailing slash from the hr element. It is not valid html and not required.

This is only in the testing sample data plugin so it is not distributed but someone might see it during development and mistakenly conclude that a trailing slash should be added

code review only

Signed-off-by: BrianTeeman <brian@teeman.net>