[5.4] Tags error message (#45569)
As with most core components the tags component will display a different error message if the alias is not unique to when the alias is not unique and trashed.
This simple PR corrects the error message for tags. This will allow a future PR to provide more helpful error messages.
Signed-off-by: BrianTeeman <brian@teeman.net>