[5.4] Error handling: Adding new shouldUseException() (#44098)
To make an easy transition from the legacy error handling with "getError" and "setError" methods, this PR introduces new methods "shouldUseExceptions" and "setUseExceptions" to the "LegacyErrorHandlingTrait". The latter allows to set a flag if exceptions should be used or not, and the former allows to check if that flag is set or not. The plan is to add this to 5.4 and then to remove all of this from 7.0 again.