[5.4] Deprecating UCM code (#44910)
The UCM system was added in Joomla 3.2.0 and already at that time had major flaws. In 12 years of development, it hasn't been extended in any way and instead we've rather reduced it instead. (The content versioning was a part of it and has been made standalone in 4.0.) I think it can be said that the concept (which barely exists to start with) has failed and we should remove the remnants of it.
This PR deprecates all UCM classes that I could find. It also removes the code where we are using the UCM classes in a functional way. For the moment it keeps it, where it might be used by third party code to keep it backwards compatible. This means that we keep this code in the batch methods, where $this->contentType and $this->type is set with this code, even though neither of the two attributes are used anywhere. We could remove this in 6.0 and drop everything in 7.0 finally.